feat: adicionado novas funções para controle junto ao banco de dados

+++ src/server/controllers/cidades/DeleteById.ts
@@ -18,6 +18,22 @@ export const deleteByIdValidation = validation((getSchema) => ({
 }));
 
 export const deleteById = async (req: Request<IParamsProps>, res: Response) => {
+  if (!req.params.id) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      errors: {
+        default: 'O parâmetro "id" precisa ser informado.',
+      },
+    });
+  }
   const result = await CidadesProvider.deleteById(Number(req.params.id));
-  return res.status(StatusCodes.OK).json(result);
+
+  if (result instanceof Error) {
+    return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+      errors: {
+        default: result.message,
+      },
+    });
+  }
+
+  return res.status(StatusCodes.NO_CONTENT).json(result);
 };
